feat: Extend Gemini context with full player tracker data
Adds 10 selectable context types for Gemini chat (labs, ultimate weapons,
modules inventory, farming/tournament presets, workshop, guardians, bots,
relics, player currencies), each backed by a dedicated ChatContext class.
Module type labels are included explicitly to prevent Gemini misclassification.

Also fixes UW stat parsing to use Double by default (handles sheet values
like "10.0") and strips "m" unit suffix in UltimateWeaponUtils.

+++ src/main/java/com/pphi/tower/model/sheets/uw/AbstractUltimateWeapon.java
@@ -24,10 +24,10 @@ public abstract class AbstractUltimateWeapon implements UltimateWeapon {
     private int stonesRequiredThree;
     private int stonesRequiredUwPlus;
 
-    protected Function<String, Number> statOneParser()   { return Integer::parseInt; }
-    protected Function<String, Number> statTwoParser()   { return Integer::parseInt; }
-    protected Function<String, Number> statThreeParser() { return Integer::parseInt; }
-    protected Function<String, Number> uwPlusStatParser() { return Integer::parseInt; }
+    protected Function<String, Number> statOneParser()    { return Double::parseDouble; }
+    protected Function<String, Number> statTwoParser()    { return Double::parseDouble; }
+    protected Function<String, Number> statThreeParser()  { return Double::parseDouble; }
+    protected Function<String, Number> uwPlusStatParser() { return Double::parseDouble; }
 
     protected AbstractUltimateWeapon(List<ValueRange> valueRanges) {
         try {
